jesper finishing relations and small details

+++ ExamProject2017/src/BE/Guild.java
@@ -3,13 +3,13 @@ package BE;
 public class Guild {
 private String GuildName;
 private int GuildId;
-private int GuildHours;
+private double GuildHours;
 private int ManagerId;
 
-public Guild(String GuildName, int GuildId, int GuildHours, int ManagerId )  {
+public Guild(String GuildName, int GuildId, double GuildHours, int ManagerId )  {
         this.GuildName = GuildName;
         this.GuildId = GuildId;
-        this.GuildHours = GuildHours;
+        this.GuildHours =  GuildHours;
         this.ManagerId = ManagerId;
     }
 public Guild(){
@@ -32,11 +32,11 @@ public Guild(){
         this.GuildId = GuildId;
     }
 
-    public int getGuildHours() {
+    public double getGuildHours() {
         return GuildHours;
     }
 
-    public void setGuildHours(int GuildHours) {
+    public void setGuildHours(double GuildHours) {
         this.GuildHours = GuildHours;
     }
